pyproject.toml
[project]
name = "unscripted-living"
version = "0.1.0"
description = "Personal retirement blog covering hobbies, activities, and thoughts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.2.0",
    "cryptography>=46.0.3",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-environ>=0.12.0",
    "gunicorn>=20.1",
    "markdown>=3.7",
    "opentelemetry-distro>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39b1",
    "opentelemetry-instrumentation-django>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "whitenoise>=6.11.0",
]
